mygcov: Update to work with gcc-5.4.0.  Remove $xfailures stuff.
The gcov output format has changed.  I do not know what version I was
using when this was originally written, nor when the format changed, or
even if it is still the same in the most recent gcc.

The previous code included reporting statistics on anomalous lines
that were classified as non-executable but were executed anyway.  I
now do not think that is needed, so removed it.