mygcov: Translate to Python using ChatGPT.
For curiosity, this commit has the text that ChatGPT produced in
response to a first prompt:

  Translate the following Perl script into Python:
  ```
  <Perl version>
  ```

Then, in response to its first output, I asked it to rewrite the
argument parsing using `argparse` (as it volunteered to do so), and
add type annotations.

I then made one change, changing the indentation from 4 spaces to 2,
and that is what is committed here.

This version of the script has not been tested at all, although I did
confirm that it is accepted by `mypy --strict`.

My intention now is to edit by hand.